FEM: examples, move meshes in sparate directory
